ai-orchestration-py/pyproject.toml
[project]
name = "ai-orchestration"
version = "0.1.0"
description = "Python AI orchestration service — LLM planning, prompt layer"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic>=2.0",
    "httpx>=0.28",
    "psycopg[binary]>=3.1",
    "psycopg-pool>=3.2",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/orchestrator"]

[tool.ruff]
target-version = "py311"
line-length = 100
